Now my life has gained its meaning since those sinful eyes behold.
Sacred land with meadows greening whose renown was often told.
This was granted me from God, to see the land and the holy sod, which in human form he trod.

Splendid lands of wealth and power, I've seen many far and near.
Yet of all are you the flower, what a wonder happened here!
That, a maid a child should bear, Lord of all the angels fair, was not this a wonder rare?

Here he was baptised the holy, that all people might be pure.
Here he died betrayed and lowly, that our bonds should not endure.
Else, our fate had been severe, hail oh cross oh thorns and spear!
Heathens woe your rage is clear!
https://www.youtube.com/watch?v=EXv8-hRB8j4